+++ Safari FIDO U2F Extension/U2FError.swift
@@ -1,0 +1,16 @@
+//
+//  U2FError.swift
+//  Safari FIDO U2F
+//
+//  Created by Sam Deane on 05/02/2017.
+//  Copyright © 2017 Yikai Zhao. All rights reserved.
+//
+
+import Foundation
+
+enum U2FError: Error {
+    case unknown(in: String)
+    case badrequest()
+    case error(u2fh_rc, in: String)
+}
+
